회원가입 관심사 오류 수정
