+++ fitchwifront/src/pages/join/components/Interest.jsx
@@ -1,4 +1,4 @@
-import React, { useCallback, useState } from "react";
+import React, { useCallback, /*useEffect,*/ useMemo, useState } from "react";
 
 import { Button, Checkbox, FormControlLabel, styled, Typography } from "@mui/material";
 import { Favorite, FavoriteBorder } from "@mui/icons-material";
@@ -31,33 +31,36 @@ export default function Interest({ joinForm, setJoinForm }) {
     setChecked({ ...checked, [e.target.name]: e.target.checked });
     console.log(e.target.name);
   };
+  const interestArr = useMemo(() => [], []);
+
+  // useEffect(() => {
+  //   console.log(interestArr);
+  // }, [interestArr]);
 
   const onCheck = useCallback(
     (e) => {
-      console.log(e.target.checked);
-
+      if (e.target.checked === true) {
+        interestArr.push(e.target.value);
+      } else {
+        interestArr.splice(interestArr.indexOf(e.target.value), 1);
+      }
       const joinObj = {
         ...joinForm,
-        memberInterest: joinForm.memberInterest,
+        memberInterest: interestArr,
       };
-      if (e.target.checked === true) {
-        joinForm.memberInterest.push(e.target.value);
-      } else {
-        joinForm.memberInterest.splice(joinForm.memberInterest.indexOf(e.target.value), 1);
-      }
       setJoinForm(joinObj);
       console.log(joinForm.memberInterest);
     },
-    [joinForm, setJoinForm]
+    [joinForm, setJoinForm, interestArr]
   );
 
   const onCheckComple = () => {
+    let interest = interestArr.join(" ");
     const joinObj = {
       ...joinForm,
-      memberInterest: joinForm.memberInterest.join(" "),
+      memberInterest: interest,
     };
     setJoinForm(joinObj);
-    console.log(setJoinForm);
   };
 
   return (
