+++ fitchwiserver/src/main/resources/application.properties
@@ -3,9 +3,9 @@ spring.datasource.driver-class-name=net.sf.log4jdbc.sql.jdbcapi.DriverSpy
 
 #spring.datasource.url=jdbc:log4jdbc:mysql://13.124.245.242:3306/kchandb?serverTimeZone=Asia/Seoul
 #spring.datasource.username=scott
-spring.datasource.url=jdbc:log4jdbc:mysql://localhost:3306/datadb?serverTimeZone=Asia/Seoul
-spring.datasource.username=dev01
-spring.datasource.password=12341234
+spring.datasource.url=jdbc:log4jdbc:mysql://localhost:3306/localfitchwi?serverTimeZone=Asia/Seoul
+spring.datasource.username=localfitchwi
+spring.datasource.password=0000
 
 #static resources
 spring.web.resources.static-locations=classpath:static/
